+++ CHANGELOG.md
@@ -10,6 +10,11 @@
   in a single task
 - MAINT: Updated SurfaceTopography to 1.11.1
 
+## 1.3.1 (2023-09-13)
+
+- BUG: Return raw ids in addition to API urls (fixes analysis download)
+- BUG: Limit list of analysis functions based on user permissions
+
 ## 1.3.0 (2023-09-11)
 
 - ENH: Updated SurfaceTopography to 1.11.0